requirements.txt
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0
numpy>=1.21.0
pandas>=1.3.0
opencv-python>=4.5.0
pillow>=8.3.0
scikit-learn>=1.0.0
python-Levenshtein>=0.12.0
editdistance>=0.5.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
pyyaml>=5.4.0
fastapi>=0.68.0
uvicorn>=0.15.0
gradio>=2.0.0
